feat(arbetsorder): hämta måtten från orderns rader till arbetsbeskrivningen
Artiklar rättas ofta EFTER att ordern skapats — men beskrivningen står då kvar
på offertens mått, och installatören bygger efter en siffra som inte gäller.
Offerten är låst vid det laget, så ordern var enda stället att komma åt det,
och där fanns ingen väg alls.

Knappen "Hämta mått från rader" finns nu i översiktens redigeringsläge, bredvid
Överlämningsnotering. Den bygger blocket ur orderns egna line_items och sparas
med resten via Spara.

Medvetet knapp och inte automatik: ordern har TVÅ spar-cykler (artiklarna mot
/line-items, översikten mot order-PATCH:en), och saveArticles kör applyWorkOrder
som läser om draften från servern. En automatik i klienten hade skrivit i ett
state som skrivs över några millisekunder senare — den hade sett ut att fungera
och tyst inte gjort något. Serverside-synk vid artikelspar är vägen om det
behövs, men det ändrar vad ett artikelspar gör och är ett eget beslut.

Orderraderna bär m2/thickness_mm/density/article_name. `construction` saknas på
rader som lagts till direkt på ordern — då faller etiketten tillbaka på
artikelnamnet, så blocket byggs ändå.

Måttlogiken flyttad från app/crm/offerter/quoteSerializers.ts till
lib/domains/crm/measurementBlock.ts: två featureytor behöver den nu, och
arbetsordern ska inte importera ur offertmappen. Testerna följer med till
tests/crm/measurementBlock.test.ts. Ny regenerateMeasurementBlock för ytor utan
automatik (alltid "bygg om"), så anropet inte behöver skicka blocket som sin
egen föregångare för att trigga borttagningen.
